Add struct p_root as a pool-allocated list
It uses the facilities already implemented in pool.c to enable fast
insertion (allocation) and removal. To leverage the benefits of using pool
allocation, an insertion position cannot be given, and the index of a
newly inserted item is not guaranteed to be 1 + the last one (as there
may be free slots in the middle of the list); instead, an optional index
argument (pointer to int) is passed in order to obtain the index.

Note: using an index number to get an item from a list of this kind incurs
greater overhead than simply dereferencing the data pointer of the
struct p_item pointer returned from item appending (allocation)
methods.