Refactored training code into environment
This is a general way to represent a "problem" where a neural network
might be applicable. It is quite flexible, its support ranging from standard
problems (like classification or regression), to "external" ones (like
a game client or database), and inbetween (like a classifier, with
annkh linked to a data processing
program).