Fix the index set by pl_allocate
poollist is meant to be a wrapper around pool to allow using
the pool allocation utility as a container of sorts.

pl_allocate was setting *index to the underlying p_root index
given by pool.c, rather than to the index of the item as
set by the pool list. This commit fixes that.