Make the weight update relative in t_evstrat.c
Weight updates will now go toward each population member's corresponding
weight, rather than simply have it be added (multiplied by a learning
rate and divided by the population count, but still). This is to ensure
that evstrats training will actually reach a desired (or move away from
an undesired) set of network weights.

Of course, as of now, training still is not functional, and debugging the
cause of that is still due.