Ignore Box-Muller pair, use one value only
boxmuller.c provides a normal distribution PRNG implementation that is, 
in turn, based on rand() (via rng.c).

The Box-Muller transform method to normally distributed random number 
generation always produces pairs. We're not interested in pairs of 
numbers, only one at a time.

Sadly, this may not be as performant as using the method once every two 
calls, and storing the other value to return it every other call. But I 
doubt this would be annkh's performance bottleneck, should one arise. :)